Switch to CircleCI

+++ builder/build.sh
@@ -102,11 +102,11 @@ umount -l ${BUILD_PATH}/sys
 
 # package image filesytem into two tarballs - one for bootfs and one for rootfs
 # ensure that there are no leftover artifacts in the pseudo filesystems
-rm -rf ${BUILD_PATH}/{dev,sys,proc}/*
+rm -rf ${BUILD_PATH:?}/{dev,sys,proc}/*
 
 tar -czf /image_with_kernel_boot.tar.gz -C ${BUILD_PATH}/boot .
 du -sh ${BUILD_PATH}/boot
-rm -Rf ${BUILD_PATH}/boot
+rm -Rf ${BUILD_PATH:?}/boot
 tar -czf /image_with_kernel_root.tar.gz -C ${BUILD_PATH} .
 du -sh ${BUILD_PATH}
 ls -alh /image_with_kernel_*.tar.gz
